Update tsconfig.json
This version includes additional compiler options such as the ECMAScript version to target, the module system to use, and enabling all strict type checking options. It also includes an "include" property that specifies which files to include in the compilation process, and an "exclude" property that specifies which files to exclude.